breaking out our value and address types into their own headers makes it a little more clear to work with.
