+++ src/CMakeLists.txt
@@ -164,6 +164,8 @@ set(edb_SRCS
 	${PROJECT_SOURCE_DIR}/include/Symbol.h
 	${PROJECT_SOURCE_DIR}/include/ThreadsModel.h
 	${PROJECT_SOURCE_DIR}/include/Types.h
+	${PROJECT_SOURCE_DIR}/include/Value.h
+	${PROJECT_SOURCE_DIR}/include/Address.h
 	${PROJECT_SOURCE_DIR}/include/Util.h
 	${PROJECT_SOURCE_DIR}/include/version.h
 )
